+++ cmd/provehito/vertical_slice_test.go
@@ -13,9 +13,9 @@ import (
 	"testing"
 	"time"
 
-	"github.com/provehito-project/provehito/core/failure"
-	"github.com/provehito-project/provehito/core/process"
-	"github.com/provehito-project/provehito/core/workspace"
+	"github.com/rremedio-web/provehito/core/failure"
+	"github.com/rremedio-web/provehito/core/process"
+	"github.com/rremedio-web/provehito/core/workspace"
 )
 
 func TestVerticalSlice(t *testing.T) {