+++ internal/releasecheck/releasecheck_test.go
@@ -13,7 +13,7 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/provehito-project/provehito/internal/releasecheck"
+	"github.com/rremedio-web/provehito/internal/releasecheck"
 )
 
 func TestCheckValidArchive(t *testing.T) {